v1/requirements.txt
pandas==2.0.0
boto3==1.28.0
python-dotenv==1.0.0